% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{_PACKAGE}
\name{aae.pop.templates}
\alias{aae.pop.templates}
\title{aae.pop.templates: parameterised population dynamics models}
\description{
aae.pop.templates includes population dynamics
  models for a range of species to work with the \code{aae.pop}
  package (described [here](https://aae-stats.github.io/aae.pop/)).

  Currently includes models for \code{\link{murray_cod}} and
  \code{\link{macquarie_perch}}.

  See \code{\link{get_template}} and \code{\link{get_args}}
  for details on defining custom population model templates.
}
\keyword{internal}
