% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\alias{simulate.template}
\title{Methods for population dynamics templates}
\usage{
\method{simulate}{template}(
  object,
  nsim = 1,
  seed = NULL,
  ...,
  init = NULL,
  options = list(),
  args = list()
)
}
\description{
Methods to simulate and plot population dynamics
  defined in \code{template} objects.
}
